bugfix - convert data to double, otherwise plot fails on single/double difference
git-svn-id: svn+ssh://svn.fcdonders.nl/home/svnroot/fieldtrip/trunk@8395 0cf7c7f0-3615-4144-b4e6-68da3bce3cd0

+++ realtime/example/ft_realtime_signalviewer.m
@@ -154,7 +154,7 @@ while true
     %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
     % put the data in a fieldtrip-like raw structure
-    data.trial{1} = dat;
+    data.trial{1} = double(dat);
     data.time{1}  = offset2time(begsample, hdr.Fs, endsample-begsample+1);
     data.label    = hdr.label(chanindx);
     data.hdr      = hdr;
